[Fixtures] Toggable trait and interface
