+++ src/Sylius/Bundle/CoreBundle/DataFixtures/Factory/ToggableTrait.php
@@ -1,0 +1,32 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Bundle\CoreBundle\DataFixtures\Factory;
+
+use Zenstruck\Foundry\ModelFactory;
+
+/**
+ * @mixin ModelFactory
+ */
+trait ToggableTrait
+{
+    public function enabled(): self
+    {
+        return $this->addState(['enabled' => true]);
+    }
+
+    public function disabled(): self
+    {
+        return $this->addState(['enabled' => false]);
+    }
+}
